+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
 ==
+## [1.1.1] - 2024-02-08
+### Modified
+- Made several UI components abstract
+### Added
+- QuitButton component
+- OnClicked abstract method on ButtonComponent
+
 ## [1.1.0] - 2024-02-01
 ### Added
 - Added support for Unity 2019.4.X versions.